clean up + add to mainstream
